+++ spec/hermod/xml_section_builder/integer_node_spec.rb
@@ -14,15 +14,15 @@ module Hermod
 
       it "should accept a valid number" do
         subject.day_of_the_week 7
-        value_of_node("DayOfTheWeek").must_equal "7"
+        expect(value_of_node("DayOfTheWeek")).must_equal "7"
       end
 
       it "should raise an error if the number is above the maximum" do
-        proc { subject.day_of_the_week 8 }.must_raise InvalidInputError
+        expect { subject.day_of_the_week 8 }.must_raise InvalidInputError
       end
 
       it "should raise an error if the number is below the minimum" do
-        proc { subject.day_of_the_week 0 }.must_raise InvalidInputError
+        expect { subject.day_of_the_week 0 }.must_raise InvalidInputError
       end
     end
   end