+++ spec/hermod_spec.rb
@@ -2,6 +2,6 @@ require 'minitest_helper'
 
 describe Hermod do
   it "should have a version number" do
-    Hermod::VERSION.wont_be_nil
+    expect(Hermod::VERSION).wont_be_nil
   end
 end